fix(composer): isolate web hover handlers
